[Modules] (Moved services)
- Moved services and their tests into appropriate modules.
