+++ internal/email/service/service.go
@@ -1,4 +1,4 @@
-package email
+package service
 
 import (
 	"bytes"
