+++ internal/user/service/service.go
@@ -1,4 +1,4 @@
-package user
+package service
 
 import (
 	"fmt"
